requirements.txt
discord.py
python-dotenv
Pillow
aiohttp
PyNaCl
